lint and move
